+++ src/extension/FixCop/BuildOutputCopAnalyzer/Entities/AnalyzedOutputLine.ts
