+++ src/extension/FixCop/BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0008.ts
