+++ src/extension/FixCop/BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0137.ts
