+++ src/extension/FixCop/BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0206.ts
