+++ src/extension/FixCop/CommandFixAssignedButUnusedVariableAA0206.ts
@@ -6,10 +6,10 @@ import { SyntaxTreeExt } from "../AL Code Outline Ext/syntaxTreeExt";
 import { TextRangeExt } from "../AL Code Outline Ext/textRangeExt";
 import { ALFullSyntaxTreeNode } from "../AL Code Outline/alFullSyntaxTreeNode";
 import { SyntaxTree } from "../AL Code Outline/syntaxTree";
-import { AnalyzerAA0008 } from "../BuildOutputCopAnalyzer/AnalyzerAA0008";
-import { AnalyzerAA0206 } from "../BuildOutputCopAnalyzer/AnalyzerAA0206";
-import { AnalyzedOutputLineAA0008 } from "../BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0008";
-import { AnalyzedOutputLineAA0206 } from "../BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0206";
+import { AnalyzerAA0008 } from "./BuildOutputCopAnalyzer/AnalyzerAA0008";
+import { AnalyzerAA0206 } from "./BuildOutputCopAnalyzer/AnalyzerAA0206";
+import { AnalyzedOutputLineAA0008 } from "./BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0008";
+import { AnalyzedOutputLineAA0206 } from "./BuildOutputCopAnalyzer/Entities/AnalyzedOutputLineAA0206";
 import { OwnConsole } from "../console";
 import { DocumentUtils } from "../Utils/documentUtils";
 import { MyTerminal } from "../Terminal/terminal";
